Make accountflags tab work again
This deletes the feature that signs transaction offline.

This is because ripple recommends setting LastLedgerSequence when
submitting transactions. It must be determined when signing the
transaction. There is no way to know what a proper value should be if we
can't talk to a rippled server.

See also https://developers.ripple.com/reliable-transaction-submission.html#lastledgersequence