Maintain a separate api instance for orderbook
Not doing so may lead to one callback called twice with the same
transaction.

This can happen because both ripple-client-desktop and
ripple-lib-orderbook subscribe to RippleAPI, one for account, the other
for transactions. Before this commit, they share the same api instance.
So when user creates an offer, the callback in ripple-lib-orderbook will
be called twice with the offer create transaction, one for each
subscription. This will cause the order be inserted twice in orderbook.

In addition, this simplifies the way that network.connect gets server
settings. On 'serverChange' event, no data is passed to callback.
Instead global variable Options.connection is used.